# Speed up muxing of GlobalState.