This is a fix for #193. To speed up the mux operation, it represents the global state as a stack of maps: One map for the top-level memory contents, and then a separate map containing the changes for each pending branch frame.

This change gives a significant speedup for subgoals in the `ecdsa-crucible.saw` example from saw-script that have a lot of conditional branches. Setup+simulation time for method `leq` goes from 18s to 0.24s (75x). Method `field_red` goes from 1.7s to 0.58s (3x). Method `group_red` goes from 1.3s to 0.42s (3x). No subproofs have any appreciable slowdown.

The drawback is that the implementation is kind of complicated. I should add some more detailed comments before I'd say it's ready to merge.