@brianhuffman 2019-04-22
About whether `ANY` types should be notified of push/abort: I thought that was a separate issue that should be addressed in a separate patch. This patch is designed to be just about performance, without changing functionality.